+++ src/file_manager.py
@@ -1,0 +1,18 @@
+import os
+import yaml
+
+
+class FileManager:
+    @staticmethod
+    def create_folder(folder_path: str) -> None:
+        os.makedirs(folder_path, exist_ok=True)
+
+    @staticmethod
+    def safe_yaml_file(file_path: str) -> dict:
+        with open(file_path) as file:
+            return yaml.safe_load(file)
+
+    @staticmethod
+    def damp_yaml_file(file_path: str, data: dict) -> None:
+        with open(file_path, "w") as file:
+            yaml.dump(data, file)
