+++ src/utils.py
@@ -1,9 +1,6 @@
-import os
 import re
 from calendar import monthrange
 from datetime import datetime
-
-import yaml
 
 
 class ConfigException(Exception):
@@ -28,19 +25,3 @@ def get_monthrange(month: int) -> int:
 
 def compile_regex(regex: str) -> re.Pattern:
     return re.compile(regex)
-
-
-class FileManager:
-    @staticmethod
-    def create_folder(folder_path: str) -> None:
-        os.makedirs(folder_path, exist_ok=True)
-
-    @staticmethod
-    def safe_yaml_file(file_path: str) -> dict:
-        with open(file_path) as file:
-            return yaml.safe_load(file)
-
-    @staticmethod
-    def damp_yaml_file(file_path: str, data: dict) -> None:
-        with open(file_path, "w") as file:
-            yaml.dump(data, file)
